Fix verl + vllm-ascend(version 0.9.1+) running failed issue
