+++ verl/workers/sharding_manager/fsdp_vllm.py
@@ -44,7 +44,7 @@ from verl.utils.fsdp_utils import (
 from verl.utils.model import check_exclude_modules, check_target_modules, convert_weight_keys
 from verl.utils.profiler import GPUMemoryLogger, log_gpu_memory_usage, simple_timer
 from verl.utils.torch_functional import check_device_is_available
-from verl.utils.vllm_utils import TensorLoRARequest, VLLMHijack, is_version_ge, patch_vllm_moe_model_weight_loader
+from verl.utils.vllm_utils import TensorLoRARequest, VLLMHijack, is_version_ge
 
 from .base import BaseShardingManager
 
@@ -328,6 +328,8 @@ class FSDPVLLMShardingManager(BaseShardingManager):
                     return k
 
                 updated_params = {replace_lora_wrapper(k): v for k, v in updated_params.items()}
+
+        from verl.utils.vllm_patch import patch_vllm_moe_model_weight_loader
 
         patch_vllm_moe_model_weight_loader(model)
         device = get_device_id()  # used when fsdp2 set cpu_offload_policy
